# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "description": "Request schema for user login.",
        "properties": {
          "identifier": {
            "description": "Email or username",
            "title": "Identifier",
            "type": "string"
          },
          "password": {
            "description": "User password",
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "identifier",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "description": "Response schema for user login.",
        "properties": {
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "JWT access token if successful",
            "title": "Access Token"
          },
          "message": {
            "description": "Response message",
            "title": "Message",
            "type": "string"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "JWT refresh token if successful",
            "title": "Refresh Token"
          },
          "success": {
            "description": "Whether login was successful",
            "title": "Success",
            "type": "boolean"
          },
          "token_type": {
            "default": "bearer",
            "description": "Token type",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "User information if successful",
            "title": "User"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
      "RefreshTokenRequest": {
        "description": "Request schema for token refresh.",
        "properties": {
          "refresh_token": {
            "description": "Valid refresh token",
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshTokenRequest",
        "type": "object"
      },
      "RefreshTokenResponse": {
        "description": "Response schema for token refresh.",
        "properties": {
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "New JWT access token if successful",
            "title": "Access Token"
          },
          "message": {
            "description": "Response message",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "description": "Whether refresh was successful",
            "title": "Success",
            "type": "boolean"
          },
          "token_type": {
            "default": "bearer",
            "description": "Token type",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "title": "RefreshTokenResponse",
        "type": "object"
      },
...
      "SignupRequest": {
        "description": "Request schema for user signup.",
        "properties": {
          "email": {
            "description": "User's email address",
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "firstname": {
            "description": "User's first name",
            "maxLength": 50,
            "minLength": 1,
            "title": "Firstname",
            "type": "string"
          },
          "lastname": {
            "description": "User's last name",
            "maxLength": 50,
            "minLength": 1,
            "title": "Lastname",
            "type": "string"
          },
          "password": {
            "description": "User password",
            "maxLength": 72,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "description": "Unique username",
            "maxLength": 30,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "email",
          "username",
          "password",
          "firstname",
          "lastname"
        ],
        "title": "SignupRequest",
        "type": "object"
      },
      "SignupResponse": {
        "description": "Response schema for user signup.",
        "properties": {
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "JWT access token if successful",
            "title": "Access Token"
          },
          "message": {
            "description": "Response message",
            "title": "Message",
            "type": "string"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "JWT refresh token if successful",
            "title": "Refresh Token"
          },
          "success": {
            "description": "Whether signup was successful",
            "title": "Success",
            "type": "boolean"
          },
          "token_type": {
            "default": "bearer",
            "description": "Token type",
            "title": "Token Type",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Created user ID if successful",
            "title": "User Id"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "title": "SignupResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "JWT": {
        "description": "JWT authentication using Bearer token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered agent hub with multi-provider LLM support",
    "title": "AgentHub API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticate a user and return JWT tokens.\n\nUsers can log in with either their email or username along with their password.\n\nArgs:\n    request: Login request containing identifier (email/username) and password\n    \nReturns:\n    LoginResponse with success status, message, tokens, and user info if successful",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Refresh an access token using a valid refresh token.\n\nThis endpoint allows clients to obtain a new access token without re-authenticating.\nThe refresh token must be valid and not expired.\n\nArgs:\n    request: Refresh token request containing the refresh token\n    \nReturns:\n    RefreshTokenResponse with new access token if successful",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Token",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/signup": {
      "post": {
        "description": "Register a new user account.\n\nThis endpoint uses the LangGraph signup workflow to validate and create a new user.\nValidation includes:\n- Email format and uniqueness\n- Username format (3-30 chars) and uniqueness\n- Password strength (8-72 chars, complexity requirements)\n- Name length (1-50 chars)\n\nArgs:\n    request: Signup request containing user details\n    \nReturns:\n    SignupResponse with success status, message, and tokens if successful",
        "operationId": "signup_api_v1_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signup",
        "tags": [
          "authentication"
        ]
      }
    },
...
  }
}